Update JavaScriptInterface.java
Changed setProgress(double) to float

+++ app/src/main/java/com/fmsys/snapdrop/JavaScriptInterface.java
@@ -273,7 +273,7 @@ public class JavaScriptInterface {
     }
     
     @JavascriptInterface
-    public void setProgress(final double progress) {
+    public void setProgress(final float progress) {
         if (progress > 0) {
             context.transfer = true;
         } else {
